Update and rename CODE.md to intro.md
The objective is to be the first player to get all three of their coloured pawns
from their start space, around the board to their "home" space. The pawns are
normally moved in a clockwise direction but can be moved backward if directed.
Movement of pawns is directed by the drawing of a card.The board game is laid
out in a square with 20 spaces per side, with each player assigned their own
coloured. Start location and Home locations offset towards the centre, one per
side. Four five-square paths, one per colour, lead from the common outer path
towards a player's Home and are designated their "Safety Zone". On each side
are two "Slides"
, grouping five or six spaces each. The player is assigned a
random move/task to perform on the board at his/ her chance.